+++ lua/neotest-golang/init.lua
@@ -6,7 +6,7 @@ local ast = require("neotest-golang.ast")
 local runspec_dir = require("neotest-golang.runspec_dir")
 local runspec_file = require("neotest-golang.runspec_file")
 local runspec_test = require("neotest-golang.runspec_test")
-local results_dir = require("neotest-golang.results_dir")
+local parse = require("neotest-golang.parse")
 
 local M = {}
 
@@ -147,13 +147,13 @@ function M.Adapter.results(spec, result, tree)
   if spec.context.pos_type == "dir" then
     -- A test command executed a directory of tests and the output/status must
     -- now be processed.
-    local results = results_dir.results(spec, result, tree)
+    local results = parse.results(spec, result, tree)
     M.workaround_neotest_issue_391(result)
     return results
   elseif spec.context.pos_type == "test" then
     -- A test command executed a single test and the output/status must now be
     -- processed.
-    local results = results_dir.results(spec, result, tree)
+    local results = parse.results(spec, result, tree)
     M.workaround_neotest_issue_391(result)
     return results
   end
